+++ lib/oh_my_flutter.dart
@@ -24,9 +24,10 @@ export 'src/extensions/date_time_extension/date_time_extension.dart' show DateTi
 export 'src/extensions/double_extension.dart' show DoubleExtension;
 export 'src/extensions/object_extension.dart' show ObjectExtension;
 export 'src/extensions/oklch_extension.dart' show OklchExtension;
+export 'src/extensions/string_extension.dart' show StringExtension;
 export 'src/extensions/velocity_extension.dart' show VelocityExtension;
 export 'src/oklch/oklch.dart' show Oklch;
-export 'src/telephony.dart' show Telephony;
+export 'src/phone_number.dart' show PhoneNumber;
 export 'src/whatsapp.dart' show Whatsapp;
 export 'src/widgets/controlled_visibility/controlled_visibility.dart' show ControlledVisibility, VisibilityController;
 export 'src/widgets/interactive_swipe_dismiss/interactive_swipe_dismiss.dart'